requirements.txt
attrs==25.3.0
cattrs==25.1.1
click==8.2.1
markdown-it-py==4.0.0
mdurl==0.1.2
Pygments==2.19.2
PyYAML==6.0.2
rich==14.1.0
shellingham==1.5.4
typer==0.16.1
typing_extensions==4.15.0
